+++ NEWS.md
@@ -1,5 +1,10 @@
 # Spectra 1.11
 
+## Changes in 1.11.6
+
+- Add `filterPrecursorPeaks` to allow filtering peaks within each spectrum
+  relative to the precursor m/z of the spectrum.
+
 ## Changes in 1.11.5
 
 - Add `scalePeaks` function (see [issue